requirements.txt
twisted
autobahn
python-nss
pytest
pytest-cov
pytest-asyncio